+++ R/adj.to.assoc.R
@@ -1,32 +1,3 @@
-#' Diagonal Part of a Matrix
-#'
-#' Similarly to upper.tri and lower.tri, returns a matrix of logicals to identify the diagonal of a square matrix
-#'
-#' @param M matrix or other R object with length(dim(x)) == 2. For back compatibility reasons, when the above is not fulfilled, as.matrix(x) is called first.
-#'
-#' @return square logical matrix with diagonal of TRUEs
-#'
-#' @examples
-#' M<- matrix(sample(1:10,16,replace = TRUE),4,4)
-#' diagonal<- function(M) {upper.tri(M,diag = TRUE)&!upper.tri(M,diag = FALSE)}
-#' diagonal(M)
-#'
-diagonal<- function(M) {upper.tri(M,diag = TRUE)&!upper.tri(M,diag = FALSE)}
-
-#' Identify coordinates of non null non diagonal elements of a matrix
-#'
-#' @param M a square matrix
-#'
-#' @return an array of row and col indices of the positive and non diagonal elements
-#'
-#' @examples
-#' M<- matrix(sample(1:10,16,replace = TRUE),4,4)
-#' diagonal<- function(M) {upper.tri(M,diag = TRUE)&!upper.tri(M,diag = FALSE)}
-#' non.zero.non.diag<- function(M) {which(M>0&!diagonal(M),arr.ind = TRUE,useNames = TRUE)}
-#' non.zero.non.diag(M)
-#'
-non.zero.non.diag<- function(M) {which(M>0&!diagonal(M),arr.ind = TRUE,useNames = TRUE)}
-
 #' Adjacency matrix to association indices matrix
 #'
 #' Calculate the association indices of choice from an adjacency matrix.