+++ NEWS.md
@@ -1,5 +1,7 @@
 # ConfiNet (development version)
 * Added focal scan simulation function simu.focal (need cleaning)
+* Pre-implementation of user-defined probabilities for simu.scan and simu.focal
+* Added option to output either list of scans or weighted adjacency matrix in simu.scan
 
 # ConfiNet 0.5.1
 * Added group scan simulation function simu.scan